remove ent (stop decoding encoded html entities)
we shouldn't actually be decoding entities - the govuk-frontend
templates return safe HTML and we should treat it as such, or we'll
decode entities that are intended to be escaped as part of the test
suite (for example ensuring that `item.text` fields are properly
escaped.

This was flagged up with v4.4.0 of govuk-frontend, which introduces a
test case where there's a `<p>` tag that contains text that includes an
escaped `<p>` tag. We were decoding the escaped text, resulting in
nested P tags which are a violation of the HTML spec, causing a parsing
error.